Added readme and rearranged folder structure
