+++ src/App.js
@@ -3,7 +3,7 @@ import { BenefitsSection } from 'components/BenefitsSection/BenefitsSection'
 import { HeaderSection } from './components/HeaderSection/HeaderSection'
 import { PickYourPositionSection } from './components/PickYourPositionSection/PickYourPosition'
 import { ReviewSection } from './components/ReviewSection/ReviewSection'
-import { QuestionSection } from './components/QuestionSection/Questions'
+import { QuestionSection } from './components/QuestionSection/QuestionSection'
 import { Footer } from './components/Footer/Footer'
 
 export const App = () => {
